model/requirements.txt
flask
transformers
SentencePiece
python-dotenv
accelerate
flask-cors
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1+cu121